+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pass `mapVariant` and `mapRevision` values in `perform`
 - **BREAKING CHANGE**: .supr files are no longer allowed
 - Optionally pass `version` to `getProfile`
+- Pass `security` and `parameters` values in `getProvider`
+
 
 ## [1.5.2] - 2022-06-15
 ### Fixed